use doubles for unpaired probs

+++ H/LPfold.h
@@ -2,7 +2,7 @@
 #define LPFOLD_H
 
 extern void  update_pf_paramsLP(int length);
-extern struct plist *pfl_fold(char *sequence, int winSize, int pairSize, 
-			      float cutoff, float *pup);
+extern struct plist *pfl_fold(char *sequence, int winSize, int pairSize,
+			      float cutoff, double *pup);
 
 #endif
